+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/Sec1EcPrivateKeyInfo.kt
@@ -5,7 +5,6 @@ package at.asitplus.awesn1.crypto
 
 import at.asitplus.awesn1.Asn1BitString
 import at.asitplus.awesn1.ObjectIdentifier
-import at.asitplus.awesn1.serialization.Asn1Tag.ConstructedBit
 import at.asitplus.awesn1.serialization.Asn1Tag
 import at.asitplus.awesn1.serialization.ExplicitlyTagged
 import kotlinx.serialization.Serializable
@@ -25,7 +24,7 @@ import kotlinx.serialization.Serializable
  * (See also [errata](https://www.rfc-editor.org/errata/rfc5915)
  */
 @Serializable
-data class EcPrivateKeyInfo(
+data class Sec1EcPrivateKeyInfo(
     val version: Int,
     val privateKey: ByteArray,
     @Asn1Tag(tagNumber = 0u)
@@ -45,7 +44,7 @@ data class EcPrivateKeyInfo(
     )
 
     override fun equals(other: Any?): Boolean =
-        other is EcPrivateKeyInfo &&
+        other is Sec1EcPrivateKeyInfo &&
             version == other.version &&
             privateKey.contentEquals(other.privateKey) &&
             parameters == other.parameters &&