+++ CHANGELOG.md
@@ -24,6 +24,7 @@
     * Rename `TbsCertificate` -> `X509TbsCertificate`
     * Rename `RsaPrivateKeyInfo` -> `Pkcs1RsaPrivateKeyInfo`
     * Rename `RsaOtherPrimeInfo` -> `Pkcs1RsaOtherPrimeInfo`
+    * Rename `EcPrivateKeyInfo` -> `Sec1EcPrivateKeyInfo`
 
 ## 0.2.1
 Equivalent to 0.2.0 but maven central is more brittle than ever so publishing 0.2.0 went south.